Sweetalert

+++ inventarios/modulos/epps/ajax-grid-data.php
@@ -70,7 +70,7 @@ while ($row = mysqli_fetch_array($query)) { // preparing an array
 	$nestedData[] = date("d/m/Y", strtotime($row["Fecha"]));
 	$nestedData[] = '<td><center>
                     <a href="editar.php?ID_Implementos=' . $row['ID_Implementos'] . '"  data-toggle="tooltip" title="Editar datos" class="btn btn-sm btn-info"> <i class="fa-solid fa-pen-to-square" style="color: #f2eded;"></i> </a>
-                    <a href="index.php?action=delete&ID_Implementos=' . $row['ID_Implementos'] . '"  data-toggle="tooltip" title="Eliminar" class="btn btn-sm btn-danger"> <i class="fa-solid fa-trash-can" style="color: #f2eded;"></i> </a>
+                    <a href="index.php?action=delete&ID_Implementos=' . $row['ID_Implementos'] . '" onclick="return confirmacion()"  data-toggle="tooltip" title="Eliminar" class="btn btn-sm btn-danger"> <i class="fa-solid fa-trash-can" style="color: #f2eded;"></i> </a>
                     </center></td>';
 
 	$data[] = $nestedData;
